first path scraping list of Phillips auctions.
used list of auction to scrape links to eaach lot that was sold, looks like site layout was changed for most of the auction, got 80 of 270 auctions, need to look at the 190 left to see how the site is structred. once all/majority of auctions are scraped need to scrape each lot.